FCT: DLS: Ajout des bits 1HZ, 2HZ et 5HZ + Ajout Grammaire Sortie Analogique
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@4408 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/json.c
@@ -70,7 +70,7 @@
 /* Entrée: le builder, le nom du parametre, la valeur                                                                         */
 /* Sortie: néant                                                                                                              */
 /******************************************************************************************************************************/
- void Json_add_double ( JsonBuilder *builder, gchar *name, gfloat valeur )
+ void Json_add_double ( JsonBuilder *builder, gchar *name, gdouble valeur )
   { json_builder_set_member_name  ( builder, name );
     json_builder_add_double_value ( builder, valeur );
   }
